srcs/userManagement/requirements.txt
asgiref==3.8.1
Django==4.2.13
sqlparse==0.5.0
typing_extensions==4.12.0
psycopg[binary]==3.1.12

djangorestframework==3.15.1.

django-cors-headers==4.3.1

pillow==10.4.0
requests==2.25.1
channels==4.1.0
daphne==4.1.2